@@ -3,12 +3,14 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@types/classnames": "^2.2.6",
     "@types/jest": "23.3.9",
     "@types/node": "10.12.2",
     "@types/react": "16.4.18",
     "@types/react-dom": "16.0.9",
-    "react": "^16.6.0",
-    "react-dom": "^16.6.0",
+    "classnames": "^2.2.6",
+    "react": "16.7.0-alpha.0",
+    "react-dom": "16.7.0-alpha.0",
     "react-scripts": "2.1.1",
     "typescript": "3.1.6"
   },
